refactor(home): extract AgentPromptList into its own file
Moves the agent-prompt pin/unpin list (AgentPromptList + PromptRow) out
of add-tile-drawer.tsx into its own module. It only reads its own
props/hooks and shares no local state with the rest of the drawer.
